Exercício 03 resolvido.

+++ Q1/main.py
@@ -13,12 +13,12 @@ class ProgressaoAritmetica:
         self.a1 = a1
         self.r = r
 
-    def somar(self) -> int:
+    def somar(self) -> float:
         # ultimo_t = primeiro_t + razão * (qnt_t - 1)
         ultimo_termo = self.a1 + self.r * (self.n - 1)
         return self.n * (self.a1 + ultimo_termo) / 2
     
-    def gerar_termos(self) -> list[int]:
+    def gerar_termos(self) -> list[float]:
         termos = [self.a1]
         value = self.a1
         for _ in range(self.n - 1):
